app/requirements.txt
fastapi[standard]==0.115.12
pymongo==4.13.0
motor==3.7.1
pytest==8.3.5
pytest-asyncio==1.0.0
pydantic-settings==2.9.1
bcrypt==4.3.0
aiobcrypt==3.2.0.post1
Faker==37.3.0
python-jose==3.4.0